pyproject.toml
[tool.poetry]
name = "catmanagementbackend"
version = "0.1.0"
description = ""
authors = ["KairzhanovMurat <murat99010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
pytest = "^8.3.3"
django-cors-headers = "^4.4.0"
drf-spectacular = "^0.27.2"
pytest-django = "^4.9.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
